refactor: rename SubscribePipeline
